backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
supabase==2.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
groq==0.4.0

# Optional: runtime.txt (to specify Python version)
python-3.11.x

# Optional: render.yaml (for infrastructure as code)
services:
  - type: web
    name: mybizsherpa-backend
    env: python
    buildCommand: pip install -r requirements.txt
    startCommand: uvicorn main:app --host 0.0.0.0 --port $PORT
    envVars:
      - key: GROQ_API_KEY
        sync: false
      - key: SUPABASE_URL
        sync: false
      - key: SUPABASE_ANON_KEY
        sync: false